// u-fe/__generated__/ts/FilterCandidates.ts
/**
 * Copyright (c) Meta Platforms, Inc. and affiliates.
 *
 * @generated SignedSource<<1e6a0b9b67b819851ca695d199d7484e>>
 */


import type { PropertyCandidates } from './PropertyCandidates.ts';

export interface FilterCandidates {
  properties: PropertyCandidates[];
  tags: string[];
  dynamic_type_keys: string[];
}